+++ CHANGES.md
@@ -11,6 +11,10 @@
 
 ## develop
 
+- [CHANGE] liwebrtc のバージョンを m140.7339.2.0 に上げる
+  - macOS, iOS が利用している clang, libc++ を Apple Clang のものから libwebrtc 管理下の Clang のものに変えたので破壊的変更となります。
+  - @melpon
+
 ## 2025.1.0
 
 **リリース日**: 2025-09-12